Primeira versão do endpoint Alertar seguidores próximos ( raio x Kms ))

+++ model/Position.go
@@ -7,7 +7,7 @@ import (
 // swagger:model
 type Position struct {
 	gorm.Model `swaggerignore:"true"`
-	Latitude   float32 `json:"Latitude" binding:"required"`
-	Longitude  float32 `json:"Longitude" binding:"required"`
+	Latitude   float64 `json:"Latitude" binding:"required"`
+	Longitude  float64 `json:"Longitude" binding:"required"`
 	UserID     uint    `json:"UserId" gorm:"not null"`
 }
